+++ agent/cmd/server/nginx_conf/stream_default.conf
@@ -1,0 +1,5 @@
+server {
+    proxy_pass mysql_cluster;
+    access_log /www/sites/domain/log/access.log streamlog;
+    error_log /www/sites/domain/log/error.log;
+}
