+++ agent/constant/website.go
@@ -4,8 +4,9 @@ const (
 	WebRunning = "Running"
 	WebStopped = "Stopped"
 
-	ProtocolHTTP  = "HTTP"
-	ProtocolHTTPS = "HTTPS"
+	ProtocolHTTP   = "HTTP"
+	ProtocolHTTPS  = "HTTPS"
+	ProtocolStream = "TCP/UDP"
 
 	NewApp       = "new"
 	InstalledApp = "installed"
@@ -15,6 +16,7 @@ const (
 	Proxy      = "proxy"
 	Runtime    = "runtime"
 	Subsite    = "subsite"
+	Stream     = "stream"
 
 	SSLExisted = "existed"
 	SSLAuto    = "auto"
