+++ agent/init/migration/migrate.go
@@ -57,6 +57,7 @@ func InitAgentDB() {
 		migrations.AddGPUMonitor,
 		migrations.UpdateDatabaseMysql,
 		migrations.InitIptablesStatus,
+		migrations.UpdateWebsite,
 	})
 	if err := m.Migrate(); err != nil {
 		global.LOG.Error(err)
