+++ agent/router/ro_website.go
@@ -93,5 +93,6 @@ func (a *WebsiteRouter) InitRouter(Router *gin.RouterGroup) {
 		websiteRouter.POST("/crosssite", baseApi.OperateCrossSiteAccess)
 
 		websiteRouter.POST("/exec/composer", baseApi.ExecComposer)
+		websiteRouter.POST("/stream/update", baseApi.UpdateStreamConfig)
 	}
 }
