+++ frontend/src/lang/modules/ko.ts
@@ -2576,6 +2576,9 @@ const message = {
 
         batchOpreate: '일괄 작업',
         batchOpreateHelper: '웹사이트를 일괄 {0}, 계속 작업하시겠습니까?',
+        stream: 'TCP/UDP 프록시',
+        streamPorts: '수신 포트',
+        streamPortsHelper: '쉼표로 구분하세요. 예: 3306, 3307',
     },
     php: {
         short_open_tag: '짧은 태그 지원',
