+++ frontend/src/views/website/website/config/basic/index.vue
@@ -1,6 +1,12 @@
 <template>
     <div :style="{ '--main-height': mainHeight + 'px' }">
-        <el-tabs tab-position="left" v-model="tabIndex" v-if="id > 0" class="custom-tabs" ref="tabsRef">
+        <el-tabs
+            tab-position="left"
+            v-model="tabIndex"
+            v-if="id > 0 && website.type != 'stream'"
+            class="custom-tabs"
+            ref="tabsRef"
+        >
             <el-tab-pane :label="$t('website.domainConfig')" name="0">
                 <Domain :key="id" :id="id" v-if="tabIndex == '0'"></Domain>
             </el-tab-pane>
@@ -59,6 +65,20 @@
                 <Other :id="id" v-if="tabIndex == '12'"></Other>
             </el-tab-pane>
         </el-tabs>
+        <el-tabs
+            tab-position="left"
+            v-model="tabIndex"
+            v-if="id > 0 && website.type == 'stream'"
+            class="custom-tabs"
+            ref="tabsRef"
+        >
+            <el-tab-pane :label="$t('website.other')" name="12">
+                <Other :id="id" v-if="tabIndex == '12'"></Other>
+            </el-tab-pane>
+            <el-tab-pane :label="$t('website.stream')" name="13">
+                <Stream :id="id" v-if="tabIndex == '13'"></Stream>
+            </el-tab-pane>
+        </el-tabs>
     </div>
 </template>
 
@@ -81,6 +101,7 @@ import PHP from './php/index.vue';
 import RealIP from './real-ip/index.vue';
 import Resource from './resource/index.vue';
 import Cors from './cors/index.vue';
+import Stream from './stream/index.vue';
 
 const props = defineProps({
     website: {
@@ -98,9 +119,10 @@ const id = computed(() => {
     return props.website.id;
 });
 const tabIndex = ref('0');
+const menuKey = ref('site-tabIndex-');
 
 watch(tabIndex, (newVal) => {
-    localStorage.setItem('site-tabIndex', newVal);
+    localStorage.setItem(menuKey.value, newVal);
 });
 
 const handleResize = () => {
@@ -123,9 +145,15 @@ const handleScroll = (event: WheelEvent) => {
 };
 
 onMounted(() => {
-    const storedTabIndex = localStorage.getItem('site-tabIndex');
+    menuKey.value = 'site-tabIndex-' + props.website.id;
+    const storedTabIndex = localStorage.getItem(menuKey.value);
     if (storedTabIndex !== null) {
         tabIndex.value = storedTabIndex;
+    } else {
+        tabIndex.value = '0';
+        if (props.website.type == 'stream') {
+            tabIndex.value = '12';
+        }
     }
     window.addEventListener('resize', handleResize);
     document.addEventListener('wheel', handleScroll, { passive: false });
