+++ frontend/src/views/website/website/config/basic/load-balance/index.vue
@@ -90,7 +90,7 @@ const loading = ref(false);
 const operateRef = ref();
 const delRef = ref();
 const fileRef = ref();
-const Algorithms = getAlgorithms();
+const Algorithms = getAlgorithms('');
 
 const buttons = [
     {
